chore: move all directives to their own files
