+++ src/lib/flexbox/flexbox-module.ts
@@ -1,20 +1,30 @@
-// RxJS Operators used by the classes...
-
 import 'rxjs/add/operator/filter';
 import 'rxjs/add/operator/map';
-
 import {ModuleWithProviders, NgModule} from '@angular/core';
-
 import {MediaQueriesModule} from '../media-query/media-query-module';
-
-import {FlexAlignDirective, FlexDirective, FlexFillDirective, FlexOffsetDirective, FlexOrderDirective} from './api/flex';
-import {LayoutAlignDirective, LayoutDirective, LayoutWrapDirective} from './api/layout';
+import {FlexDirective} from './api/flex';
+import {LayoutDirective} from './api/layout';
 import {HideDirective, ShowDirective} from './api/show-hide';
 import {MediaQueryAdapter} from './media-query/media-query-adapter';
+import {FlexAlignDirective} from './api/flex-align';
+import {FlexFillDirective} from './api/flex-fill';
+import {FlexOffsetDirective} from './api/flex-offset';
+import {FlexOrderDirective} from './api/flex-order';
+import {LayoutAlignDirective} from './api/layout-align';
+import {LayoutWrapDirective} from './api/layout-wrap';
+
 
 const ALL_DIRECTIVES = [
-  LayoutDirective, LayoutWrapDirective, LayoutAlignDirective, FlexDirective, FlexOrderDirective,
-  FlexOffsetDirective, FlexFillDirective, FlexAlignDirective, ShowDirective, HideDirective
+  LayoutDirective,
+  LayoutWrapDirective,
+  LayoutAlignDirective,
+  FlexDirective,
+  FlexOrderDirective,
+  FlexOffsetDirective,
+  FlexFillDirective,
+  FlexAlignDirective,
+  ShowDirective,
+  HideDirective,
 ];
 
 /**
